requirements.txt
torch
numpy
PyYAML
opencv-python
matplotlib
tqdm
safetensors
